Don't check again jQuery but constructors
We use the constructor to differentiate if it is a object we know we can
stringify. Since the jQuery type of is object we can't use that.
When

we come to a object which we don't know try to copy it's properties.
That the closes we can get.

When we try to stringify a reference to the a dom element if will become
'{}' but if you try to stringify the global document 💥, therefor
replace it with "document reference"